# Improve Guid parsing performance